Keep vLLM request-shaping policy out of the neutral connector base
